Changed the structure of the pagaes and routing adding only one dinamic page
